Enhance port configuration to support TX multi-segments and improve timestamp polling logic; remove unused timestamp fields and add statistics for received packets

+++ apps/testbench/m_strings.h
@@ -273,7 +273,7 @@ static __inline__ char *m_strtolower(char *str)
 	if (!str)
 		return NULL;
 
-	for (int i = 0; i <= (int)(strlen(str)); i++)
+	for (size_t i = 0; i <= strlen(str); i++)
 		str[i] = tolower(str[i]);
 
 	return str;
@@ -293,7 +293,7 @@ static __inline__ char *m_strtoupper(char *str)
 	if (!str)
 		return NULL;
 
-	for (int i = 0; i <= (int)(strlen(str)); i++)
+	for (size_t i = 0; i <= strlen(str); i++)
 		str[i] = toupper(str[i]);
 
 	return str;
